requirements.txt
numpy==1.25.2
pandas==1.3.3
matplotlib==3.4.3
torch==2.2.2
torchvision==0.17.2
vendi-score==0.0.3
mlflow==2.21.2
IPython==8.5.0
huggingface-hub==0.16.4
tensorflow==2.16.2
pypng==0.20220715.0
tensorflow-text==2.16.1
tensorflow-datasets==4.9.2
tensorflow-hub==0.16.1